Promote the fully validated mingw-7.2-integration line onto the canonical mingw-7.2 branch.

This merge carries the exact upstream Redis 7.2.14 baseline, the Windows/MinGW portability updates, IOCP and QFork fixes, Redis 7.2 Module API support, package/service hardening, and the completed validation ledger. The integration tip was synchronized with origin before promotion; no post-7.2.14 upstream production commits are omitted.

A clean promoted-commit build, full source test matrix, package/runtime/service verification, persistence soak, and final artifact audits will be completed before creating v7.2.14.

+++ deps/lua/src/lua_struct.c
@@ -138,7 +138,7 @@ static int gettoalign (size_t len, Header *h, int opt, size_t size) {
   if (size == 0 || opt == 'c') return 0;
   if (size > (size_t)h->align)
     size = h->align;  /* respect max. alignment */
-  return (int)((size - (len & (size - 1))) & (size - 1));
+  return (size - (len & (size - 1))) & (size - 1);
 }
 
 
@@ -223,7 +223,7 @@ static int b_pack (lua_State *L) {
     switch (opt) {
       case 'b': case 'B': case 'h': case 'H':
       case 'l': case 'L': case 'T': case 'i': case 'I': {  /* integer types */
-        putinteger(L, &b, arg++, h.endian, (int)size);
+        putinteger(L, &b, arg++, h.endian, size);
         break;
       }
       case 'x': {
@@ -232,13 +232,13 @@ static int b_pack (lua_State *L) {
       }
       case 'f': {
         float f = (float)luaL_checknumber(L, arg++);
-        correctbytes((char *)&f, (int)size, h.endian);
+        correctbytes((char *)&f, size, h.endian);
         luaL_addlstring(&b, (char *)&f, size);
         break;
       }
       case 'd': {
         double d = luaL_checknumber(L, arg++);
-        correctbytes((char *)&d, (int)size, h.endian);
+        correctbytes((char *)&d, size, h.endian);
         luaL_addlstring(&b, (char *)&d, size);
         break;
       }
